feat: 将定时检查服务接口健康移动至单独结构体
