+++ src/core/mod.rs
@@ -2,7 +2,7 @@ pub mod collector;
 pub mod doctor;
 pub mod ent;
 pub mod logger;
-pub use collector::listen;
+pub use collector::{listen, ServiceCaller};
 pub use doctor::*;
 pub use ent::*;
 pub use logger::*;
