+++ src/main.rs
@@ -14,11 +14,9 @@
 //! ```
 
 mod core;
-use std::time::{SystemTime, UNIX_EPOCH};
 
 use crate::core::*;
 use tokio::sync::mpsc;
-use tokio::time;
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
 
 #[tokio::main]
@@ -44,11 +42,11 @@ async fn main() {
     //2. 启动用于监听节点状态和服务状态的任务
     tokio::spawn(async move {
         // Init Monitor
-        let mut monitor = Logger::new(dc1);
+        let mut logger = Logger::new(dc1);
         println!("into watch");
         loop {
             tokio::select! {
-                Some(event)=rx1.recv()=> monitor.log(event),
+                Some(event)=rx1.recv()=> logger.log(event),
                 else => { break }
             };
         }
@@ -56,48 +54,9 @@ async fn main() {
     });
     //3. 启动用于轮询各服务Health接口的任务
     //   同时此任务负责定时通知Monitor遍历节点以检查有哪些节点超时未更新
-    let mut services = Vec::new();
-    services.push(Service {
-        name: "rust website".to_string(),
-        api: "https://www.rust-lang.org".to_string(),
-        latency: 0,
-        last_updated: 0,
-    });
     tokio::spawn(async move {
-        let mut should_export = 0;
-        //TODO: Graceful Shutdown
-        loop {
-            println!("linglingling");
-            time::sleep(time::Duration::from_secs(10)).await;
-            for srv in &services {
-                println!("service = {:?}", srv);
-                let (status, msg) = dc2.check_service(&srv.api).await;
-                println!("check result = {:?} {:?}", status, msg);
-                tx2.send(Event::Heartbeat(HealthInfo {
-                    target: Target::Service(
-                        String::from(&srv.name),
-                        Some(Service {
-                            name: String::from(&srv.name),
-                            api: String::from(&srv.api),
-                            latency: srv.latency,
-                            last_updated: SystemTime::now()
-                                .duration_since(UNIX_EPOCH)
-                                .unwrap()
-                                .as_secs(),
-                        }),
-                    ),
-                    status: status,
-                }))
-                .await
-                .unwrap();
-            }
-            // 每轮询5次 触发一次全局汇报
-            should_export += 1;
-            if should_export == 5 {
-                tx2.send(Event::CheckAll).await.unwrap();
-                should_export = 0;
-            }
-        }
+        let srv_caller = ServiceCaller::new(dc2, tx2);
+        srv_caller.turn_on().await;
     });
 
     //4. 启动监听节点健康状况的服务
